refactor: test pool claim
